requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
decorator==4.4.2
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
filelock==3.20.3
Flask==2.3.3
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-groq==1.1.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
motor==3.6.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.1
openai==2.15.0
opencv-python-headless==4.13.0.90
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
primp==0.15.0
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.4
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.38.6
streamlit==1.53.0
sympy==1.14.0
tavily-python==0.7.19
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.32.0
watchdog==6.0.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.2.4
zstandard==0.25.0
